🐞 Fix(readme): correct and supplement README.md
I don't think subheaders should contain leading numbers like 1.1 as
it'd be a great pain for us if a subheader is removed or inserted.

The official link of Git is 'https://git-scm.com/' instead of
'https://desktop.github.com/', which is of GitHub Desktop.
Git is a version control system (VCS), while GitHub is a developer
platform using Git software. Subversion (SVN), another VCS, should be
compared to Git rather than GitHub.

The previous link of the Git installation tutorial did not point to
the tutorial itself. Instead, it pointed to the search results of Bing.
Now it is replaced with a link to a super-detailed tutorial on CSDN.

IMHO, the steps for cloning our repo are not supposed to share the same
subheader with the links to stuffs about Git.

To build the environment, I guess it is as significant as on Windows
to illustrate what to do on Linux. Most servers run Linux OS.

Thank you for considering the corrections above.