🐛 fix(server.py): make solve POST return a dict
The front-end seems to need a dict (JSON serializable object) with
properties 'x' and 'y', representing x and y values correspondingly.
My bad I'd not have proposed to return a CSV file to the front-end.
A PNG image or a CSV text file is not easy to cope with using js.
Moreover, all relative paths are made PWD-independent.
As branch iw17 was left behind, the bug on deg of poly is fixed again.